qp-trie: support big-endian as well
The unportable hack used there should be fine on any sane little-endian.
I didn't want to spend much time on it, so if little-endian can't be
inferred, a slightly less memory-efficient but portable code is used.
To be sure, in debug mode there are checks that the unportable code is OK.